NABTO-1547 fix slow closing
If we can only reach this code when the client has called close on the
stream then it should be ok to send a fin packet whenever (tcb->xmitSeq
== tcb->xmitFirst).

xmitSeq is the next sequence number to when adding data to the stream.
xmitFirst is the first unacked sequence number, aka the number. Aka when
tcb->xmitSeq == tcb->xmitFirst then there is no further data from this
side which needs to be acked from the other side.

ST_FIN_WAIT_1
ST_CLOSING,

in these states the fin initiator has called close so no more data is
added to the stream aka xmitSeq is constant until fin is sent.

in ST_LAST_ACK, the fin responder has called close on the stream meaning
no more data will be written to the stream from that end, so xmitSeq
is constant until the fin has been sent.